♻️  Refactor user into a custom hook
