+++ src/hooks/useUser.js
@@ -1,0 +1,25 @@
+import { useState, useEffect } from 'react'
+
+import { Auth } from 'aws-amplify'
+
+const useUser = () => {
+  const [user, setUser] = useState(null)
+
+  useEffect(() => {
+    checkUser()
+  }, [])
+
+  async function checkUser() {
+    try {
+      const data = await Auth.currentUserPoolUser()
+      const userInfo = { username: data.username, ...data.attributes }
+      setUser(userInfo)
+    } catch (err) {
+      console.log('error: ', err)
+    }
+  }
+
+  return { user }
+}
+
+export default useUser
