+++ src/pages/Profile.js
@@ -1,5 +1,3 @@
-import React, { useState, useEffect } from 'react'
-import { Auth } from 'aws-amplify'
 import { Dashboard } from '../layout/Dashboard'
 
 import protectedRoute from './protectedRoute'
@@ -7,22 +5,10 @@ import protectedRoute from './protectedRoute'
 import { Members } from '../components/profile/Members'
 import { Games } from '../components/profile/Games'
 
+import useUser from '../hooks/useUser'
 
 function Profile() {
-  useEffect(() => {
-    checkUser()
-  }, [])
-
-  const [user, setUser] = useState(null)
-  async function checkUser() {
-    try {
-      const data = await Auth.currentUserPoolUser()
-      const userInfo = { username: data.username, ...data.attributes }
-      setUser(userInfo)
-    } catch (err) {
-      console.log('error: ', err)
-    }
-  }
+  const { user } = useUser()
 
   return (
     <Dashboard>
@@ -51,7 +37,6 @@ function Profile() {
           {/* Games */}
           <Games />
         </div>
-
       )}
     </Dashboard>
   )
